+++ python/openinference-semantic-conventions/tests/openinference/semconv/test_attributes.py
@@ -218,7 +218,11 @@ class TestMessageContentAttributes:
         attributes = _get_attributes(MessageContentAttributes)
         assert _nested_dict(attributes) == {
             "message_content": {
+                "data": MessageContentAttributes.MESSAGE_CONTENT_DATA,
+                "encrypted_content": MessageContentAttributes.MESSAGE_CONTENT_ENCRYPTED_CONTENT,
+                "id": MessageContentAttributes.MESSAGE_CONTENT_ID,
                 "image": MessageContentAttributes.MESSAGE_CONTENT_IMAGE,
+                "signature": MessageContentAttributes.MESSAGE_CONTENT_SIGNATURE,
                 "text": MessageContentAttributes.MESSAGE_CONTENT_TEXT,
                 "type": MessageContentAttributes.MESSAGE_CONTENT_TYPE,
             }
@@ -314,6 +318,7 @@ class TestToolCallAttributes:
                     "name": ToolCallAttributes.TOOL_CALL_FUNCTION_NAME,
                 },
                 "id": ToolCallAttributes.TOOL_CALL_ID,
+                "reasoning_signature": ToolCallAttributes.TOOL_CALL_REASONING_SIGNATURE,
             },
         }